Fix: "close" was not being correctly parsed
It was the existence of the word "closet" that caused this bug. The reason the code was designed this way was so that e.g. "escape pod" would be recognized instead of being short-circuited to "escape". The fix is to require the point beyond which the longer word differs to be a space.