Fix bug related to AGI strings being ovewritten with the empty string
There was an off-by-one error in the per-string size of the AGI string variable array, such that cSetString() would make the next-numbered string empty, when what it actually meant to do was make sure the current-numbered string is NUL-terminated if it has the maximum length.

Notably, in Space Quest, the character's chosen name is stored in strings[1]. Presumably strings[0] is used somewhere; storing any value in it would have overwritten strings[1] with an empty string.