[internal] Improve implementation of "close" fix
This accomplishes the same fix as 35284aa27c4d7faebbe4080043f8b21a1a124aec, but more concisely/efficiently. This version of the fix was suggested by andersnm.

The bug was that if a word existed that was identical to an entered word except for having one extra letter at the end, it would be recognized as the longer word. The effect would be as if the command had that longer word overwritten into it in-place, followed by a space. Examples:
* "close door " would be interpreted as "closet oor ". The error message would show the originally entered command, though: `I don't understand "close door "`
* "buckle belt " in the SQ1 escape pod would be interpreted as "buckle belts", which works because "belts" is a synonym for "belt" in the game. But then, only once per session of being inside the escape pod, it'd give the error `I don't understand "belt "`. The reason for this part of it is not known yet.