romgui: Partially fix updating of status line
Previously, the status line updates were completely invisible; it was intended to show "Adding Game:" followed by the name of each game. Perhaps in 2004, computers were slow enough that this actually mattered? I'm skeptical of even that, but certainly now, the status line changes whip by instantly.

In any case, this commit creates a new bug - the status line updates overwite each other. So this is currently relegated to a branch.