[internal] Revert an accidentally committed source code change
This was accidentally included in 03b8bb2acb3d7c11a38ab7f3a2915e5ea4aed997, albeit in the form in which "mode || " is commented out (so the commit has no effect on the compiled code).
It was intended only to be a local test, in which uncommenting "mode || " would force all words to be included in the verb/noun lists of the "Enter Input" dialog in "More..." mode.