romgui, gbarom: Fix usage of vocab.bin to categorize verbs and non-verbs
It was using the wrong method of comparing strings (comparing only up to the first space), thus if "don't" is a verb, "don't touch" would be matched with "don't" in vocab.bin and classified as a verb, ignoring the actual entry for "don't touch" in vocab.bin.
This fixes the bug noticed in 890541171dcc5e000778dccd0e5a05e893c0d8ec.

There was one way in which that method of comparing strings was correct, though: it made "look around", "look under", etc., all get their verb status from "look". That will need to be done explictly for each variant of "look" now.

Related changes:
* [internal] Comment out a call to FindWordx() that had no effect