Fix word count returned by ParseInput()
The value of "inpos" was not updated by ParseInput(), meaning it would retain its previous value from the "Enter Input" dialog (i.e. "inpos" counted the number of words entered in that dialog). This resulted in the following bugs:
* The AGI parse() command would not update the word count, meaning that a subsequent said() test would think the number of words parsed was the same as whatever command last entered in "Enter Input".
  - Note that in Space Quest 1, the computer console interfaced to the cartridge retrieval unit would still accept the password "ASTRAL BODY", but only because the only command that gets there is "look screen" (or some synonym combo thereof) which has the same word count as "astral body".
* If words with spaces were entered in "Enter Input" as separate words, for example "look" "across" instead of "look across", the word count would be wrong (in this example, it'd be 2 when it should be 1).

Even with this bug, however, pressing BTN_L to recall the last command – triggering ExecuteInputDialog(FALSE) – would finally correctly update "inpos", so for example if you had entered "look" "across", it would be recalled as "look across", which BTN_B would delete in one step. After this fix, this is still the behavior (correctly).